Create SimpleTaskbar
This Bat file makes your taskbar in Windows 10 look more clean. It removes the people's icon, removes the search box and stops anything from showing up when you right click the task bar. This is useful if you want to deploy a locked down image of Windows 10 in an organisation.